handle relationship user & contact  schema and manage percular users contacts

+++ controller/contactController.js
@@ -3,17 +3,17 @@ const Contact = require('../models/contactModel');
 const express = require('express');
 //@des get all contact
 //router GET /api/contacts
-// Access Public
+// Access Private
 
 const getContacts =asyncHandler( async(req,res)=>{
-    const contacts =await Contact.find()
+    const contacts =await Contact.find({user_id: req.user.id});
     res.status(200).json(contacts);
 });
 
 
 //@des create contact
 //router POSt /api/contacts
-// Access Public
+// Access Private
 
 const createContact = asyncHandler( async(req,res)=>{
     console.log("The request is : ",req.body);
@@ -26,14 +26,15 @@ const createContact = asyncHandler( async(req,res)=>{
     const contact = Contact.create({
         name,
         email,
-        phone
+        phone,
+        user_id:req.user.id
     })
     res.status(201).json(contact);
 });
 
 //@des get contact
 //router GET /api/contacts/:id
-// Access Public
+// Access Private
 
 const getContact =asyncHandler(async(req,res)=>{
 
@@ -47,7 +48,7 @@ const getContact =asyncHandler(async(req,res)=>{
 
 //@des uppdate contact
 //router PUT /api/contacts/:id
-// Access Public
+// Access Private
 
 const updateContact =asyncHandler( async(req,res)=>{
 
@@ -55,6 +56,10 @@ const updateContact =asyncHandler( async(req,res)=>{
     if(!contact){
         res.status(404);
         throw new Error("Contact Not Found");
+    }
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("Users dont have permission to  update other users contacts");
     }
 
     const updatedContact = await Contact.findByIdAndUpdate(
@@ -76,6 +81,12 @@ const deleteContact =asyncHandler(async(req,res)=>{
         res.status(404);
         throw new Error("Contact not Found");
     }
+
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("Users dont have permission to  delete other users contacts");
+    }
+
    await Contact.findByIdAndDelete(req.params.id);
     res.status(200).json(contact);
 });
